requirements.txt
streamlit
pandas
faker
google-generativeai
python-dotenv
watchdog
